+++ sdk/src/orders/index.ts
@@ -74,6 +74,8 @@ export interface UnsignedOrder {
     deadline: number;
 }
 
+export * from "./erc7683";
+
 type ValidatedOrderVenue = OrderVenue & {
     reactor: Address;
     permit2: Address;